Swaps redirect for basepath and moves app to /
