+++ next.config.js
@@ -3,16 +3,8 @@ const { i18n } = require('./next-i18next.config')
 module.exports = {
   // The Claim Tracker application must live at /claimstatus because IDM Webgate 
   // expects the "context root" to be /claimstatus, and all other pages must be sub-paths.
-  async redirects() {
-    return [
-      {
-        source: '/',
-        destination: '/claimstatus',
-        permanent: true,
-        locale: false,
-      },
-    ]
-  },
+  basePath: '/claimstatus',
+  assetPrefix: '/claimstatus',
   i18n,
   webpack: (config, { buildId, dev, isServer, defaultLoaders, webpack }) => {
     config.node = {
